backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.23.2
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3
supabase==2.0.2
openai==1.3.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
redis==5.0.1
celery==5.3.4
sendgrid==6.10.0
twilio==8.10.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx>=0.24.0,<0.25.0